Cache Init function added

+++ src/utils/cache/sync.cache.utils.ts
@@ -4,8 +4,9 @@ import { SyncCacheDataType, syncCacheSchema, SyncErrorDataType } from "../../sch
 import { RequestActions } from "../../schemas/configs/actions.app.config.schema";
 import { ClientConfigType, SyncrhonousClientConfigDataType } from "../../schemas/configs/client.config.schema";
 import { getConfig } from "../config.utils";
-import { DBClient } from "../mongo.temp.utils";
-import { MongoUtils } from "../mongo.utils";
+import { DBClient } from "../mongo.utils";
+
+const syncCacheCollectionName = "sync_cache";
 
 export class SyncCache {
     public static getInstance(): SyncCache {
@@ -28,7 +29,17 @@ export class SyncCache {
         this.dbClient = new DBClient(syncCacheConfig.mongoURL);
     }
 
-    public async init(message_id: string, action: RequestActions) {
+    public async initialize() {
+        if(getConfig().client.type != ClientConfigType.synchronous){
+            throw new Exception(ExceptionType.SyncCache_InvalidUse, "Sync Cache should be used only in case the client connection mode is synchronous.", 400);
+        }
+
+        if (!this.dbClient.isConnected) {
+            await this.dbClient.connect();
+        }
+    }
+
+    public async initCache(message_id: string, action: RequestActions) {
         if(getConfig().client.type != ClientConfigType.synchronous){
             throw new Exception(ExceptionType.SyncCache_InvalidUse, "Sync Cache should be used only in case the client connection mode is synchronous.", 400);
         }
@@ -37,7 +48,7 @@ export class SyncCache {
             await this.dbClient.connect();
         }
 
-        const collection = this.dbClient.getDB().collection("sync_cache");
+        const collection = this.dbClient.getDB().collection(syncCacheCollectionName);
         await collection.insertOne({
             message_id,
             action,
@@ -55,7 +66,7 @@ export class SyncCache {
             await this.dbClient.connect();
         }
 
-        const collection = this.dbClient.getDB().collection("sync_cache");
+        const collection = this.dbClient.getDB().collection(syncCacheCollectionName);
         await collection.updateOne({
             message_id,
             action
@@ -75,7 +86,7 @@ export class SyncCache {
             await this.dbClient.connect();
         }
 
-        const collection = this.dbClient.getDB().collection("sync_cache");
+        const collection = this.dbClient.getDB().collection(syncCacheCollectionName);
         await collection.updateOne({
             message_id,
             action
@@ -93,7 +104,7 @@ export class SyncCache {
             await this.dbClient.connect();
         }
 
-        const collection = this.dbClient.getDB().collection("sync_cache");
+        const collection = this.dbClient.getDB().collection(syncCacheCollectionName);
         const result = await collection.findOne({
             message_id,
             action
@@ -103,11 +114,35 @@ export class SyncCache {
             return null;
         }
 
+        return syncCacheSchema.parse(result);
+    }
+
+    public async deleteData(message_id: string, action: RequestActions) {
+        if(getConfig().client.type != ClientConfigType.synchronous){
+            throw new Exception(ExceptionType.SyncCache_InvalidUse, "Sync Cache should be used only in case the client connection mode is synchronous.", 400);
+        }
+
+        if (!this.dbClient.isConnected) {
+            await this.dbClient.connect();
+        }
+
+        const collection = this.dbClient.getDB().collection(syncCacheCollectionName);
         await collection.deleteOne({
             message_id,
             action
         });
-        
-        return syncCacheSchema.parse(result);
+    }
+
+    public async clear() {
+        if(getConfig().client.type != ClientConfigType.synchronous){
+            throw new Exception(ExceptionType.SyncCache_InvalidUse, "Sync Cache should be used only in case the client connection mode is synchronous.", 400);
+        }
+
+        if (!this.dbClient.isConnected) {
+            await this.dbClient.connect();
+        }
+
+        const collection = this.dbClient.getDB().collection(syncCacheCollectionName);
+        await collection.deleteMany({});
     }
 };
